doll is moving in script
The mannequin is moving based on the position of key markers on the keypoints. However, it is only connected to the wrists and ankles, which is fine for development.
Next is to have it move on an offset, so that you can have the mannequin where ever.
Also fixed  keypoints